requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorize==1.1.0
Django==2.2
djangorestframework==3.9.3
factory-boy==2.11.1
Faker==1.0.5
idna==2.8
PTable==0.9.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.24.2

